pyproject.toml
[project]
name = "intelligent-multi-agent-tutor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.24.1",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
]
